refactor(canon): keep legacy helper wrapper scoped
